AG-15850 Fix `lint:depcruise` error
