+++ packages/ag-charts-types/src/api/activeState.ts
@@ -1,0 +1,17 @@
+export interface AgActiveItemState {
+    /** Where the item activation originates from. */
+    type: 'series-area' | 'legend';
+    /** The unique identifier of the series that this picked datum belongs to. */
+    seriesId: string;
+    /** The unique identifier of the picked datum. */
+    itemId: string | number;
+}
+
+export interface AgActiveState {
+    /**
+     * The active series datum shape. If the entire series is active, then `itemId` will be set to `undefined`.
+     */
+    activeItem?: AgActiveItemState;
+    /** The frozen state. When the picked item is frozen, user interactions with the chart will be ignored and not updated the currently picked item. */
+    frozen: boolean;
+}
