+++ packages/ag-charts-types/src/api/initialStateOptions.ts
@@ -3,6 +3,7 @@ import type { AgInitialStateLegendOptions } from '../chart/legendOptions';
 import type { Ratio } from '../chart/types';
 import type { AgAutoScaledAxes } from '../chart/zoomOptions';
 import type { AgPriceVolumeChartType } from '../presets/financial/priceVolumeOptions';
+import type { AgActiveState } from './activeState';
 import type { AgStateSerializableDate } from './stateTypes';
 
 // Theme
@@ -47,22 +48,4 @@ export interface AgInitialStateZoomRange {
 export interface AgInitialStateZoomRatio {
     start?: Ratio;
     end?: Ratio;
-}
-
-export interface AgActiveItemState {
-    /** Where the item activation originates from. */
-    type: 'series-area' | 'legend';
-    /** The unique identifier of the series that this picked datum belongs to. */
-    seriesId: string;
-    /** The unique identifier of the picked datum. */
-    itemId: string | number;
-}
-
-export interface AgActiveState {
-    /**
-     * The active series datum shape. If the entire series is active, then `itemId` will be set to `undefined`.
-     */
-    activeItem?: AgActiveItemState;
-    /** The frozen state. When the picked item is frozen, user interactions with the chart will be ignored and not updated the currently picked item. */
-    frozen: boolean;
 }
